src/app/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
databricks-sdk>=0.39.0
pydantic>=2.9.0
